+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add migrate script to pull to new version of deepgnn.
 
+- Add debug mode to MultiWorkersConverter, using debug=True will now disable multiprocessing and show error messages.
+
 ### Changed
 - Remove FeatureType enum, replace with np.dtype. FeatureType.BINARY -> np.uint8, FeatureType.FLOAT -> np.float32, FeatureType.INT64 -> np.int64.